Add antarctic_ice_shelves_polys to download
This merge switches to using a python script to do the downloading.
This let us get the version number more easily from cartopy, as
well as add features that are not downloaded by default.

I have kind of cheated by downloading the python file we want and
then importing it as a module.  If a better way exists, I'm
eager to hear it!